[#80130] 측정소 별 실시간 측정정보 API 데이터  삽입 완료, 일부 로직 수정 필요

+++ src/main/java/com/example/realpilot/excelModel/RegionData.java
@@ -18,7 +18,7 @@ public class RegionData {
     private String createdDate;
     private Integer gridX;
     private Integer gridY;
-    private Float tmX;
-    private Float tmY;*/
+    private Double tmX;
+    private Double tmY;*/
 
 }
